Add compatability for redis clusters in redis check
The RedisHealthCheck faield when using a redis cluster and phpredis
due to the requirement of an argument to the ping method on the
RedisCluster class (found in the redis extension).

This change adds compatability by detecting which redis connection
we are using, and then handling the ping appropriately for either
a cluster or a single node.

The tests for RedisHealthcheck have been reworked to ensure that
the correct flow is taken dependent on configuration and connection
type being used.