Fix failing tests on php 8
The failing tests were due to the change in later versions of
laravel. The default redis client changed from predis to phpredis.
This meant that the tests failed for predis, as the phpredis
client was set.

To fix this I have explicitly set the client in the config for
each test. This seemed to be the simplest way to ensure it works
correctly accross all laravel versions.